+++ src/components/EditorComponent.vue
